+++ src/main/java/yoon/community/dto/comment/CommentReadCondition.java
@@ -1,6 +1,5 @@
 package yoon.community.dto.comment;
 
-import io.swagger.annotations.ApiOperation;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -14,5 +13,5 @@ import javax.validation.constraints.PositiveOrZero;
 public class CommentReadCondition {
     @NotNull(message = "게시글 번호를 입력해주세요.")
     @PositiveOrZero(message = "올바른 게시글 번호를 입력해주세요. (0 이상)")
-    private Integer boardId;
+    private Long boardId;
 }
