+++ src/main/java/yoon/community/dto/report/BoardReportRequest.java
@@ -14,10 +14,9 @@ import javax.validation.constraints.NotNull;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BoardReportRequest {
-
     @ApiModelProperty(value = "신고 당하는 게시글 아이디", notes = "신고 당할 게시글 아이디를 입력해주세요.", required = true, example = "3")
     @NotNull(message = "신고할 게시글의 아이디를 입력해주세요.")
-    private int reportedBoardId;
+    private Long reportedBoardId;
 
     @ApiModelProperty(value = "신고 사유", notes = "신고 사유를 입력해주세요.", required = true, example = "Insulting")
     @NotBlank(message = "신고 사유를 입력하세요. ")
